+++ include/actl/graph/edge_list.hpp
@@ -59,7 +59,7 @@ public:
 
     explicit edge_list_edges() = default;
 
-    int edge_count() const { return edges_.size(); }
+    int edge_count() const { return static_cast<int>(edges_.size()); }
 
     template <class... Ts>
     std::pair<edge, bool> try_add_edge(vertex u, vertex v, Ts&&... args) {
